ovl: invalidate snapshot dentry if lower was renamed
With "redirect_dir=nofollow" we allow lower dir to be renamed.
We need to invalidate the overlay dentry in that case, so lookup will
be called to reassemble the lowerstack.

To be absolute certain that lower dentry was not renamed we would need to
know the redirect path that lead to it, but that is not necessary.
Instead, we just store the hash of the parent/name from when we composed
the stack, which gives a good enough probablity to detect a lower rename
and is much less complexity.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>