ovl: do not return error on remote dentry cache expiry
ovl_dentry_revalidate() calls d_revalidate() on remote fs dentry and
returns -ESTALE (user visible error) if remote dentry cache is invalid.
The official reason for that behavior is because changes to underlying
fs are not expected and particularly changes on the remote server which
invalidate the local dentry cache.

However, network filesystem may choose return zero from d_revalidate()
due to cache timeout expiry or for any other reason for that matter,
even if the underlying remote files where not changed.

For a mounted network filesystem client, this dentry revalidation event
will be contained within vfs, but with overlayfs on top of the network
filesystem, an error will be exposed to user.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>