ovl: create a whiteout in change tracking snapshot before lower create
Before creating a new lower object, if the overlay snapshot dentry
is not a whiteout or positive, explicitly whiteout the overlay snapshot
dentry to indicate that a new object or in fact a new tree may have been
created under that whiteout.

If snapshot upper dir and lower dir are on the same filesystem and
that filesystem metadata is journaled and strictly ordered, there
is no need to fsync the snapshot upper dir before proceeding with
create of lower object.  For other setups, the 'dirsync' mount option
for the overlay snapshot mount can be used to request explicit fsync
of the whiteout marker before lower object create.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>