fsnotify: send MOVE_SELF event with parent/name info
MOVE_SELF event does not get reported to a parent watching children
when a child is moved, but it can be reported to sb/mount mark or to
the moved inode itself with parent/name info if group is interested
in parent/name info.

Use the fsnotify_parent() helper to send a MOVE_SELF event and adjust
fsnotify() to handle the case of an event "on child" that should not
be sent to the watching parent's inode mark.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>