ovl: do not follow mismatch origin in overlay snapshot
With an overlay snapshot, lower dir can be deleted and a new dir created
in its place, so do not follow to lower dir when origin xattr exists and
does not match lower dir.

Do not print "failed to verify origin..." warnings with overlay snapshot
because those verification failures are expected when lower objects are
replaced.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>