ovl: add mark with ignored mask after marked in snapshot
Add inode mark on lowerdir with ignored mask.
If lowerdir is copied to snapshot ignore only FS_PRE_MODIFY.
If lowerdir or an ancestor is opaque in snapshot, ignore also
FS_PRE_MODIFY_NAME, because children do not need to and cannot be
marked in snapshot.

The mark is added with flag FSNOTIFY_ADD_MARK_NO_IREF, because we
do not want to hog system resources with those optimization marks.
When inodes are evicted those marks are destroyed and on the next
modification, the overlay watch event handler will lookup the snapshot
again and re-instate the mark with ignored mask.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>