ovl: invalidate overlay snapshot dentry if lower was renamed
With overlay snapshot mount we expect lower objects to be renamed.
We need to invalidate the overlay dentry in that case, so lookup
will be called to reassemble the lowerstack.

To be absolute certain that lower dentry was not renamed we would need to
store the parent and name that were used in lookup, but we do not need
this level of certainty.

Instead, we just store the hash of the parent/name from when we composed
the stack, which gives a good enough probablity to detect a lower rename
and with much less complexity.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>