ovl: add mount option "watch"
With mount option "watch", the overlay acts as a "persistent change
tracking snapshot" of a single lower tree.  Any change to the lower tree
while overlay snapshot is mounted, will copy the lower object's ancestry
to upper dir before the change.  Before creating a new lower object, if
overlay has no entry in the upper tree, a whiteout will be created in the
upper tree.

An alternative way to configure an overlay snapshot is by replacing the
"lowerdir=" mount option with "watchdir=".  For example, in the following
mount commands:

mount -t overlay snapshot $S \
      -o lowerdir=$L,upperdir=$U,workdir=$W,watch

mount -t overlay snapshot $S \
      -o watchdir=$L,upperdir=$U,workdir=$W

$L is the path of the watched tree.
$U contains the "persistent change tracking snapshot".
$S is a read-only overlay snapshot mount.

The content of the overlay snapshot mount is not well defined and is not
meant to be used for change tracking, but as a sanity check, with cold
dentry cache, lower entries that were created after overlay snapshot was
mounted should not be visible in directory tree iteration of the overlay
snapshot mount.

This option requires CONFIG_OVERLAY_FS_WATCH and that a single lower
path is provided along with upperdir and workdir.

This change only sets up a mark on the lower fs to get FS_PRE_MODIFY
events when user specifies the "watch" mount option.
Later changes will take care of triggering copy up in overlay snapshot on
modifications to lower layer.

The overlay snapshot mount is forced read-only, so it cannot be used for
rename of upper entries and cannot create redirects.  This implies that
lookup in the change tracking tree is more efficient by name and without
inodes index.  The metacopy feature is irrelevant because the change
tracking tree does not contain any data at all.

Therefore, overlay snapshot mount disables the features: redirect follow,
metacopy and inode index.  If any of those features are explicitly
requested by mount options along with "watch", the mount will fail.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>