ovl: set mismatch origin opaque with redirect_dir=origin
With mount option "redirect_dir=origin", if lower dir does not match the
origin fh stored in upper dir, or if no lower dir was found by name,
set the upper dir opaque.

This mount option implies this is an overlay snapshot mount with all
the properties that come with it (e.g. forced r/o mount).

Signed-off-by: Amir Goldstein <amir73il@gmail.com>