ovl: sync directory after copy to metacopy snapshot
With regular overlay copy up, fsync is done in ovl_copy_up_data().
With metacopy snapshot copy up, only an empty file or its parent dir are
created as change markers.

We need to make sure that the change markers are persistent on-disk
before making modifications to make sure that after a crash the change
markers will be found.

We request copy up to fsync the upper dir before marking the overlay
dentry copied up, so that a parallel thread cannot continue to make
changes before copy up is safe on-disk.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>