fanotify: introduce filesystem view mark
A filesystem view is a subtree of a filesystem accessible from a specific
mount point.  When marking an FS view, user expects to get events on all
inodes that are accessible from the marked mount, even if the events
were generated from another mount.

In particular, the events such as FAN_CREATE, FAN_MOVE, FAN_DELETE that
are not delivered to a mount mark can be delivered to an FS view mark.

One example of a filesystem view is btrfs subvolume, which cannot be
marked with a regular filesystem mark.

Another example of a filesystem view is a bind mount, not on the root of
the filesystem, such as the bind mounts used for containers.

A filesystem view mark is composed of a heads sb mark and an sb_view mark.
The filesystem view mark is connected to the head sb mark and the head
sb mark is connected to the sb object. The mask of the head sb mask is
a cumulative mask of all the associated sb_view mark masks.

Filesystem view marks cannot co-exist with a regular filesystem mark on
the same filesystem.

When an event is generated on the head sb mark, fsnotify iterates the
list of associated sb_view marks and filter events that happen outside
of the sb_view mount's root.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>