fanotify: introduce FAN_MARK_IN_USERNS
WIP

TODO: limit to single sb, limit to sb, limit to non init user ns

A filesystem in-userns mark is intended to request events on all the
operations performed on a filesystem via an idmapped mount.

When marking a filesystem in-userns, user expects to get events on
filesystem operations via all mounts that are decendants of the marked
idmapped mount, even if the events were generated by a process from
another userns.

In particular, the events such as FAN_CREATE, FAN_MOVE, FAN_DELETE that
are not delivered to a mount mark can be delivered to a filesystem
in-userns mark.

A common use case for of a filesystem in-userns marks is idmapped mounts
used as home directories for containers.

A filesystem in-userns mark is connected to the mnt_userns object and has
a weak refernce to the sb of the idmapped mounted filesystem.

Every group can have just one filesystem in-userns mark per userns.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>