fanotify: introduce filesystem subtree mark
A filesystem subtree mark is intended to request events on all the
inodes in the same filesystem that are decendants of a subtree root
directory.  At this time a filesystem subtree mark can only be set
on mount points.

When marking a filesystem subtree, user expects to get events on all
inodes that are decendants of the marked mount root, even if the events
were generated from another mount.

In particular, the events such as FAN_CREATE, FAN_MOVE, FAN_DELETE that
are not delivered to a mount mark can be delivered to a filesystem
subtree mark.

A common use case for of a filesystem subtree is a bind mount, not on
the root of the filesystem, such as the bind mounts used for containers.

A filesystem subtree mark is composed of an internal sb mark and a
subtree mark.  The filesystem subtree mark is connected to the internal
sb mark and the internal sb mark is connected to the sb object.
The mask of the internal sb mark is a cumulative mask of all the
associated subtree mark masks.

When an event matches the internal sb mark cumulative mask of a group,
fsnotify iterates the list of associated subtree marks and checks if the
event object is in any subtree that is interested in the event.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>