fanotify: introduce FAN_MARK_IDMAPPED
WIP - TODO: cleanup userns_sb marks when last idmapped mount goes away

Setting up mount and filesystem marks is allowed when user is capable
in the user ns that mounted the filesystem.

When the fanotify_mark() flag FAN_MARK_IDMAPPED is provided along with
FAN_MARK_MOUNT or FAN_MARK_FILESYSTEM, a user that is capable in the
user ns of an idmapped mount is allowed to setup mount and filesystem
marks.

In case of an idmapped filesystem mark, the group will receive events
only when the filesystem operations where executed via an idmapped
mount that uses the same userns as the idmapped mount of the marked path.

Specifically, the group will receive event for filesystem operations
executed via the mount of the marked path, any clones of that mount and
any children bind mounts of that mount.

This means that if an idmapped mount was used to "carve out" a section
of the host filesystem to a container (e.g. as home directory), that
container user will be able to setup a mark and receieve events on all
the operations performed on that "carved out" filesystem section within
its own container as well as from children containers that it may spawn.

Note that the events such as FAN_CREATE, FAN_MOVE, FAN_DELETE that
cannot be delivered to a group using a mount mark can be delivered to a
group using an idmapped filesystem mark.

Unlike filesystem marks that are connected to an sb object, an idmapped
filesystem mark is connected to the userns object and has a weak refernce
to the sb of the idmapped mounted filesystem.

Currently, every group can have just one idmapped filesystem mark per
userns, so if multiple filesystems are idmapped using the same userns,
separate group will be needed to monitor each idmapped filesystem.
This limitation could be lifted in the future.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>