ovl: add mark with ignored mask for lowerdir not under lower root
Add inode mark on lowerdir with ignored mask for lowerdirs of no
interest.

The mark is added with flag FSNOTIFY_ADD_MARK_NO_IREF, because we
do not want to hog system resources with those optimization marks.
When inodes are evicted those marks are destroyed and on the next
modification, the overlay watch event handler will lookup the index
again and re-instate the mark with ignored mask.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>