ovl: create index entry in change tracking snapshot before modify
On pre modify events, if lower dir is not yet indexed, create the index
entry for that lower dir.

A directory index entry in the overlay index dir indicates that there
may be a modification somewhere in the lower tree under that entry.

We need to fsync indexdir before allowing lower data modification,
because data writes can be flushed before the index changes are
committed to disk.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>