ovl: create index entries in change tracking snapshot before modify
On modify permission events, if lower dir is not yet indexed, create
index entries for all the directories in ancestry.

A directory index entry in the overlay index dir indicates that there
may be a modification somewhere in the lower tree under that entry.

We need to fsync indexdir before allowing lower data modification,
because data writes can be flushed before the index changes are
committed to disk.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>