ovl: whiteout a directory index entry before lower rename
WIP

Before moving lower directory, ...

if the overlay snapshot dentry
is an upper directory whose lower dir was deleted or renamed, explicitly
set the upper snapshot directory opaque to avoid missing changes to the
children of a renamed lower directory.

This is needed to fix "rename and rename back" case demonstrated by
the unionmount test:
 ./run --sn --verify --recycle rename-pop-dir

Where a subdir that is created in a renamed directory is not marked
as a change in snapshot after its parent is renamed back to under its
original copied up upper.  The opaque directory serves as a change marker
that indicates that the entires tree under it should be considered new.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>