ovl: implement lookup in change tracking snapshot index
Snapshot index lookup is about finding out if a lower directory has
already been recorded in the change tracking snapshot.  This is done
in a similar way of decoding a lower nfs file handle of an overlay
mount.

Before modification of lower layer path, we look for an index entry for
the lower dir in index dir.

When looking up in overlay snapshot index, what we need to know about the
overlay index entry is if it is positive/negative/whiteout so we know if
the lower object change has already been recorded.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>