ovl: add mount option "watch"
With mount option "watch", overlayfs watches filesystem changes on a
single lower tree.  Any change to the lower tree while overlayfs is
mounted, will record the lower object's ancestry in index dir before
the change.

The option "watch" requires CONFIG_OVERLAY_FS_WATCH and that a single
lowerdir path is provided on the same fs as upperdir and workdir.
The option "watch=mnt" can be used to watch a lower mount that is not
on the same fs as upperdir.

The "watch" feature along with the "index" feature enables indexing of
directories, which is usually only enabled with "nfs_export" feature.

This change adds marks on the lower filesystem to get move and modify
permission events when user specifies the "watch" mount option.

The permission event handler prevents modification to the lower layer
when overlay is mounted with "index=off".

Later changes will take care of recording changes in the index on
modifications to lower layer and allowing the recorded modifications.

If a directory is copied up before lower modification, the index already
has a valid entry for the lower directory and lower object modification
will not require any new record.

If on the other hand a lower modification is made before lower directory
is copied up, an attempt to lookup and/or copy up the lower direcory
will fail, because the existing index entry does not have a valid "upper"
reference to a copied up upper dir.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>