ovl: add ignored mark for indexed lower dir
After lowerdir is recorded in index we have no need for getting
more pre modification events on that lowerdir, so add a mark with
ignored mask in that case.

We still need to get pre move events for indexed lowerdir in order to
index moved subdirs, so exclude pre move event from the ignored mask.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>