ovl: create index entries for all the directories in ancestry
On pre modify events, if lower dir is not yet indexed, create the index
entries for all the directories in ancestry before creating the index
entry for the lowerdir itself.

Unless directories are moved from outside of the lower layer into the
lower layer or moved not through the watched lower mount, a lower dir
is not expected to be indexed unless all of its ancestors upto and
including the lower layer root dir are already indexed.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>