fs: allow open by file handle inside userns
open_by_handle_at(2) requires CAP_DAC_READ_SEARCH in init userns,
where most filesystems are mounted.

Relax the requirement to allow a user with CAP_DAC_READ_SEARCH
inside userns to open by file handle in filesystems that were
mounted inside that userns.

In addition, also allow open by handle in an idmapped mount, which is
mapped to the userns while verifying that the returned open file path
is under the root of the idmapped mount.

This is going to be needed for setting an fanotify mark on a filesystem
and watching events inside userns.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>