fanotify: support sb mark filtered by idmapped mount
For a group created inside userns, adding a mount mark is allowed
if the mount is idmapped inside the group's userns and adding a
filesystem mark is allowed if the filesystem was mounted inside the
group's userns.

Allow also adding a filesystem mark on a filesystem that is mounted
in the group's userns via an idmapped mount.

In that case, the mark is created with an internal flag, which indicates
that events should be sent to the group only if they happened via an
idmapped mount inside the group's userns.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>