fanotify: add suport for new event FAN_LINK
Add event FAN_LINK for reporting that an inode was linked at path.

This event is reported in addition to the existing FAN_CREATE event.
The former is from the point of view of the directory where the entry
to link the inode was created.  The latter is from the point of view of
the inode being linked.

With a filesystem or mount mark the new event can be used to inform the
listener of a new object in the filesystem, which the listener can then
start watching more closely.

The event is created using the new hook fsnotify_link_at().
Unlike fsnotify_link(), this hook is called for any type of inode linked
via any syscall, not only link().  This hook also carries the mount
context so unlike FAN_CREATE, the FAN_LINK event can be added to a mask
of a mount mark.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>