ovl: annotate st_dev of unmodified vs. new and modified lower dirs
In an overlayfs watch mount, unmodified lower directories
(i.e. not indexed and not new) use the real lower fs st_dev/st_ino
unlike modified and new directories that use the overlay st_dev,
so 'du -x' on the overlayfs mount will list only the modified and the
new lower directories.

Renamed lower directories are a special case, because they return
-ESTALE on lookup.  This is needed for path indexers to know that
the entire subtree needs to be re-scanned.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>