fsnotify: enable reporting xattr events to mount marks
Pass the mount context to fsnotify_xattr() hooks from syscalls and nfsd,
so that attrib events may be added to a mount mark mask.

The fsnotify_xattr() hooks are moved out of the vfs_setxattr()
and vfsremovexattr() helpers and called without the inode lock.

The helpers vfs_setxattr_notify() and vfs_removexattr_notify() are
used to call the vfs helpers and fsnotify hook.

Some call sites that manipulate subsystem internal xattr, such as
cachefiles object xattrs and some overlayfs xattr operations, still call
the vfs helper without the fsnotify hook, because the subsystem internal
xattrs should not be the concern of a watching user.

Cc: Christian Brauner <christian.brauner@ubuntu.com>
Cc: Tyler Hicks <code@tyhicks.com>
Cc: J. Bruce Fields <bfields@fieldses.org>
Cc: Miklos Szeredi <miklos@szeredi.hu>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>