fsnotify: enable reporting create and delete events to mount marks
Add the ability to subscribe to FAN_CREATE and FAN_DELETE events on a
mount mark.

This functionality was requested by Christian Brauner to replace the use
of recursive inotify watches for detecting when some path was created
under an idmapped mount without having to monitor create events on the
entire filesystem.

An optional mnt argument is added to the fsnotify create/delete hooks.
If NULL, events are reported to inode and sb marks as before.  If mnt
context is available, then event may also be reported to mount marks.

The existing fsnotify hooks for create and delete events are moved out
from the vfs helpers, where mount context is not available. The fsnotify
hooks are explicitly called after the vfs helpers with or without mount
context.

vfs_xxx_notify() convenience wrappers are used to bundle the calls to
vfs helper and fsnotify hook from syscalls, nfsd and selective callers.

Suggested-by: Jan Kara <jack@suse.cz>
Link: https://lore.kernel.org/linux-fsdevel/20210318143140.jxycfn3fpqntq34z@wittgenstein/
Link: https://lore.kernel.org/linux-fsdevel/CAOQ4uxi-UhF=6eaxhybvdBX-L5qYx_uEuu-eCiiUzJPvz2U8aw@mail.gmail.com/
Cc: Christian Brauner <christian.brauner@ubuntu.com>
Cc: Tyler Hicks <code@tyhicks.com>
Cc: David Howells <dhowells@redhat.com>
Cc: J. Bruce Fields <bfields@fieldses.org>
Cc: Miklos Szeredi <miklos@szeredi.hu>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>