fanotify: introduce group flag FAN_REPORT_CHILD_FID
FAN_REPORT_FID is ambigious in that it reports the fid of the child for
some events and the fid of the parent for create/delete/move events.

The new flag FAN_REPORT_CHILD_FID requires FAN_REPORT_DIR_FID and is
mutually exclusive with FAN_REPORT_FID.

With flag combination FAN_REPORT_DIR_FID|FAN_REPORT_FID, in addition to
reporting FAN_EVENT_INFO_TYPE_DFID info record on events on non-dir,
the fid of the child is reported with info type FAN_EVENT_INFO_TYPE_FID.

With FAN_REPORT_CHILD_FID, the additional FAN_EVENT_INFO_TYPE_FID info
record with child fid is reported also on create/delete/move events,
regardless if the child is a subdir or not.

As a by-product of the extra child fid info, event listeners that
initialize a group with flags FAN_REPORT_DFID_NAME|FAN_REPORT_CHILD_FID,
can use the child fid in a similar manner that the inotify rename cookie
is used to pair up matching moved_{from,to} events.

To simplify the semantics of pairing moved_{from,to} events, the
create/delete/move events are never merged in a group with flag
FAN_REPORT_CHILD_FID, so due to vfs locking rules those events will
never be re-ordered in this mode.

There would not be much gain in merging those events because there cannot
be multiple create/delete/move events with the same child fid, except
for the corner case of renaming a file over and over through a small
subset of names.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>